Connect tile placement with resource manager

+++ Assets/_Scripts/Tile.cs
@@ -2,6 +2,6 @@ using UnityEngine;
 
 public class Tile :MonoBehaviour
 {
-    public float oxygen;
-    public float temperature;
+    public int oxygen;
+    public int temperature;
 }
